auth_service/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
opentelemetry-api==1.23.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
protobuf==4.25.3
psycopg2==2.9.6
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.4.0
python-dotenv==1.0.1
redis==4.4.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
starlette==0.35.1
thrift==0.16.0
typer==0.9.0
typing_extensions==4.10.0
uvicorn==0.27.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
